Added new Enemies
New enemies have been added.
The Game1.cs class has been modified so that instead of a single
texture, it accepts an array of texture. It is passed to the WaveManager
class whose constructor has been modified to accept an array of 2d
textures instead of a single texture.
Every other wave is a different enemy.
The stats have to still be modified depending on the type of enemy.
A further improvement is having both types of enemies come
simultaneously in a wave.